backend/requirements.txt
Pillow
django
djangorestframework